+++ src/controls/legend.js
@@ -555,6 +555,7 @@ const Legend = function Legend(options = {}) {
     },
     getuseGroupIndication() { return useGroupIndication; },
     getOverlaysCollapse() { return overlaysCmp.overlaysCollapse; },
+    getOverlays() { return overlaysCmp; },
     setVisibleLayersViewActive,
     addButtonToTools(button, buttonGroup) {
       if (buttonGroup === 'addLayerButton') {
