+++ src/controls/legend/overlayproperties.js
@@ -5,19 +5,33 @@ import Style from '../../style';
 const OverlayProperties = function OverlayProperties(options = {}) {
   const {
     cls: clsOptions = '',
+    group,
     layer,
     viewer
   } = options;
-
-  const cls = `${clsOptions} item`.trim();
-  const title = layer.get('title') || '';
-  const abstract = layer.get('abstract') || '';
-  const opacity = layer.getOpacity();
-  const opacityControl = layer.get('opacityControl') !== false;
-  const style = viewer.getStyle(layer.get('styleName'));
-  const legendComponent = Legend({ styleRules: style, opacity, layer, viewer, clickable: false });
-  const stylePicker = viewer.getLayerStylePicker(layer);
-
+  let cls = '';
+  let title;
+  let abstract;
+  let opacity;
+  let opacityControl;
+  let style;
+  let legendComponent;
+  let stylePicker = [];
+  if (layer) {
+    cls = `${clsOptions} item`.trim();
+    title = layer.get('title') || '';
+    abstract = layer.get('abstract') || '';
+    opacity = layer.getOpacity();
+    opacityControl = layer.get('opacityControl') !== false;
+    style = viewer.getStyle(layer.get('styleName'));
+    legendComponent = Legend({ styleRules: style, opacity, layer, viewer, clickable: false });
+    stylePicker = viewer.getLayerStylePicker(layer);
+  } else if (group) {
+    title = group.title;
+    abstract = group.description;
+    opacity = group.opacity || 1;
+    opacityControl = group.opacityControl !== false;
+  }
   let styleSelection;
   let overlayEl;
   let sliderEl;
@@ -52,8 +66,11 @@ const OverlayProperties = function OverlayProperties(options = {}) {
   }
 
   function renderStyleSelection() {
-    const html = `<div class="o-stylepicker-header text-small padding-small">Välj stil</div>${styleSelection.render()}`;
-    return hasStylePicker() ? html : '';
+    if (hasStylePicker()) {
+      const html = `<div class="o-stylepicker-header text-small padding-small">Välj stil</div>${styleSelection.render()}`;
+      return html;
+    }
+    return '';
   }
 
   function getStyleDisplayName(styleName) {
@@ -119,31 +136,38 @@ const OverlayProperties = function OverlayProperties(options = {}) {
 
   return Component({
     onInit() {
-      styleSelection = Dropdown({
-        direction: 'up',
-        cls: 'o-stylepicker text-black flex',
-        contentCls: 'bg-grey-lighter text-smaller rounded',
-        contentStyle: 'max-height: 12em; overflow-y: auto;',
-        buttonCls: 'bg-white border text-black box-shadow',
-        buttonTextCls: 'text-smaller',
-        text: getStyleDisplayName(layer.get('styleName')),
-        buttonIconCls: 'black',
-        ariaLabel: 'Välj stil'
-      });
-      const components = [legendComponent];
+      const components = [];
+      if (layer) {
+        components.push(legendComponent);
+      }
       if (opacityControl) {
         components.push(transparencySlider);
       }
       if (hasStylePicker()) {
+        styleSelection = Dropdown({
+          direction: 'up',
+          cls: 'o-stylepicker text-black flex',
+          contentCls: 'bg-grey-lighter text-smaller rounded',
+          contentStyle: 'max-height: 12em; overflow-y: auto;',
+          buttonCls: 'bg-white border text-black box-shadow',
+          buttonTextCls: 'text-smaller',
+          text: getStyleDisplayName(layer.get('styleName')),
+          buttonIconCls: 'black',
+          ariaLabel: 'Välj stil'
+        });
         components.push(styleSelection);
       }
       this.addComponents(components);
-      this.on('click', (e) => {
-        extendedLegendZoom(e);
-      });
+      if (layer) {
+        this.on('click', (e) => {
+          extendedLegendZoom(e);
+        });
+      }
     },
     onRender() {
-      viewer.getControlByName('legend').dispatch('renderOverlayProperties', { cmp: this, layer });
+      if (layer) {
+        viewer.getControlByName('legend').dispatch('renderOverlayProperties', { cmp: this, layer });
+      }
       this.dispatch('render');
       overlayEl = document.getElementById(this.getId());
       overlayEl.addEventListener('click', (e) => {
@@ -153,11 +177,25 @@ const OverlayProperties = function OverlayProperties(options = {}) {
         sliderEl = document.getElementById(transparencySlider.getId());
         sliderEl.nextElementSibling.value *= 100;
         sliderEl.addEventListener('input', () => {
-          layer.setOpacity(sliderEl.valueAsNumber);
+          if (group) {
+            group.opacity = sliderEl.valueAsNumber;
+            viewer.getLayersByProperty('group', group.name).forEach(l => l.setOpacity(sliderEl.valueAsNumber));
+            viewer.getLayersByProperty('id', `grouplayer-${group.name}`).forEach(l => l.setOpacity(sliderEl.valueAsNumber));
+          }
+          if (layer) {
+            layer.setOpacity(sliderEl.valueAsNumber);
+          }
           sliderEl.nextElementSibling.value *= 100;
         });
         sliderEl.addEventListener('change', () => {
-          layer.setOpacity(sliderEl.valueAsNumber);
+          if (group) {
+            group.opacity = sliderEl.valueAsNumber;
+            viewer.getLayersByProperty('group', group.name).forEach(l => l.setOpacity(sliderEl.valueAsNumber));
+            viewer.getLayersByProperty('id', `grouplayer-${group.name}`).forEach(l => l.setOpacity(sliderEl.valueAsNumber));
+          }
+          if (layer) {
+            layer.setOpacity(sliderEl.valueAsNumber);
+          }
         });
       }
       if (hasStylePicker()) {
@@ -171,7 +209,7 @@ const OverlayProperties = function OverlayProperties(options = {}) {
     render() {
       return `<div id="${this.getId()}" class="${cls} border-bottom">
                 <div class="padding-small">
-                  ${legendComponent.render()}
+                  ${legendComponent ? legendComponent.render() : ''}
                   ${renderStyleSelection()}
                   ${opacityControl ? transparencySlider.render() : ''}
                 </div>
